Add quizId to /edit

+++ src/routes/delete/[id]/+server.ts
@@ -10,13 +10,21 @@ export const DELETE: RequestHandler = async ({ locals: { supabase, getSession },
 		throw error(404, "Quiz ID not found");
 	}
 
-	const { data, error: err } = await supabase
+	const { data, error: err } = await supabase.from("quiz").select("user_id").eq("id", params.id);
+
+	if (err) throw error(500, "Server Error");
+	if (!data) throw error(404, "Not found");
+
+	if (data[0].user_id !== session.user.id) {
+		throw error(401, "Unauthorized");
+	}
+
+	const { data: deleteData, error: deleteErr } = await supabase
 		.from("quiz")
 		.delete()
-		.eq("user_id", session.user.id)
 		.eq("id", params.id);
 
-	if (err) {
+	if (deleteErr) {
 		throw error(500, "Server Error");
 	}
